+++ src/main/java/cart/controller/ProductApiController.java
@@ -1,8 +1,8 @@
 package cart.controller;
 
 import cart.service.ProductService;
-import cart.controller.dto.ProductRequest;
-import cart.controller.dto.ProductResponse;
+import cart.controller.dto.request.ProductRequest;
+import cart.controller.dto.response.ProductResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
