+++ src/main/java/cart/controller/dto/request/ProductRequest.java
@@ -1,4 +1,4 @@
-package cart.controller.dto;
+package cart.controller.dto.request;
 
 public class ProductRequest {
     private String name;
